Move planning api to global api
