+++ src/api/index.ts
@@ -3,3 +3,4 @@ export * from './types';
 export * from './farmer';
 export * from './customer';
 export * from './manager';
+export * from './planning';
