+++ src/modules/planning/pages/PlanningScreen/components/index.ts
@@ -1,2 +1,3 @@
 export * from './PlanningTable';
 export * from './PlanningStatisticsSection';
+export * from './BalanceSection';
