--- src/modules/planning/pages/components/BalanceSection/index.ts
@@ -1,1 +1,0 @@
-export * from './BalanceSection';
